+++ doc/history.md
@@ -77,6 +77,40 @@ Response Body (Failed):
 }
 ```
 
+## Get History (done)
+
+Endpoint: GET /api/histories/1
+
+Headers:
+
+- Authorization: token
+
+Response Body (Success):
+
+```json
+{
+    "data": [
+        {
+            "historyId": 1,
+            "disease": disease.name,
+            "description_disease": disease.description,
+            "face_img": "*.jpg",
+            "solution_acne": disease.solution,
+            "product_solution": product.name
+        },
+    ]
+}
+```
+
+Response Body (Failed):
+
+```json
+{
+  "status": 404,
+  "message": "Unauthorized"
+}
+```
+
 ## Delete History (done)
 
 Endpoint: DELETE /api/histories/:historyId